Avoid code duplication by moving the Finite State Machine implementation into it's own module. Both v3 and v5 of the BIRO solver use the same FSM.
