+++ setup.py
@@ -98,6 +98,8 @@ setup(name='montblanc',
         'montblanc.examples',
         'montblanc.impl',
         'montblanc.impl.biro',
+        'montblanc.impl.biro.common',
+        'montblanc.impl.biro.common.fsm',
         'montblanc.impl.biro.v2',
         'montblanc.impl.biro.v2.gpu',
         'montblanc.impl.biro.v2.cpu',
